[KubeAPIServerOpenIDConnect] Prefactor: extract defaulting into dedicated file
