+++ pkg/apis/settings/v1alpha1/defaults.go
@@ -42,14 +42,3 @@ func SetDefaults_ClusterOpenIDConnectPreset(obj *ClusterOpenIDConnectPreset) {
 	}
 	setDefaultServerSpec(&obj.Spec.Server)
 }
-
-func setDefaultServerSpec(spec *KubeAPIServerOpenIDConnect) {
-	if len(spec.SigningAlgs) == 0 {
-		spec.SigningAlgs = []string{DefaultSignAlg}
-	}
-
-	if spec.UsernameClaim == nil {
-		usernameClaim := DefaultUsernameClaim
-		spec.UsernameClaim = &usernameClaim
-	}
-}
