+++ pkg/apis/settings/v1alpha1/defaults_kubeapi_openidconnect.go
@@ -1,0 +1,12 @@
+package v1alpha1
+
+func setDefaultServerSpec(spec *KubeAPIServerOpenIDConnect) {
+	if len(spec.SigningAlgs) == 0 {
+		spec.SigningAlgs = []string{DefaultSignAlg}
+	}
+
+	if spec.UsernameClaim == nil {
+		usernameClaim := DefaultUsernameClaim
+		spec.UsernameClaim = &usernameClaim
+	}
+}
